Replace serial create with serialport, which has been updated much more recently

+++ espflash/src/cli/clap.rs
@@ -6,7 +6,7 @@ pub struct ConnectArgs {
     pub serial: Option<String>,
     /// Baud rate at which to flash target device
     #[clap(long)]
-    pub speed: Option<usize>,
+    pub speed: Option<u32>,
 }
 
 #[derive(Parser)]
